feat: show time to read on every bio instance
